requirements.txt
certifi==2018.1.18
chardet==3.0.4
dj-database-url==0.4.2
Django==2.0.2
django-fe-core==0.3.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
idna==2.6
mysqlclient==1.3.12
PyJWT==1.5.3
pytz==2018.3
requests==2.18.4
urllib3==1.22
